refactor(list): use component files instead of array of enabled routes
